refactor: share short and long performance workloads
